move MiddleCard, MinCard, add href for link for search and MiddleCard, fix bug
