+++ src/components/Catalogy/Search/Search.module.scss
@@ -127,6 +127,7 @@
 	& svg {
 		width: 28px;
 		margin: 0px;
+		cursor: default;
 	}
 }
