+++ src/components/Catalogy/cardCatalogy/MinCard/MinCard.module.scss
