+++ src/components/Catalogy/cardCatalogy/MinCard/MinCard.tsx
